:heavy_plus_sign: (feat): Add express as a dependency
In order to use GitHub's OAuth authentication (ideal for logging in, I
think), we need to spawn a local web server during the OAuth flow to
catch the tokens provided by GitHub after the user accepts our necessary
permissions. This is janky but as far as I know it's the only way to do
it.