:sparkles: (feat): Add login command and initial implementation
For the first pass at an implementation, we're going to instruct the
user to create their own Personal Access Token and pass it to spoon for
the application to save and use. This is not an ideal user experience,
but it will provide a much better experience than what we have right
now, while also preventing us from banging our head against the wall
getting OAuth to function in the CLI.

We still need to ensure that the application as a whole can access the
token when any command is run before merging this, or any,
implementation of the command.